# api-docs/openapi.json
{"openapi":"3.0.0","info":{"title":"Dorar.net Hadith API","version":"1.0.0","description":"API for searching and retrieving hadith from Dorar.net"},"servers":[{"url":"http://localhost:5000/v1"}],"paths":{"/api/hadith/search":{"get":{"summary":"Search for hadiths using Dorar.net API","parameters":[{"in":"query","name":"value","schema":{"type":"string"},"required":true,"description":"Search query for hadith text"},{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number for pagination"},{"in":"query","name":"removehtml","schema":{"type":"boolean"},"description":"Remove HTML tags from results"},{"in":"query","name":"specialist","schema":{"type":"boolean"},"description":"Include specialist hadiths"},{"in":"query","name":"xclude","schema":{"type":"string"},"description":"Words or phrases to exclude from search"},{"in":"query","name":"st","schema":{"type":"string","enum":["w","a","p"]},"description":"Search type (w=all words, a=any word, p=exact phrase)"},{"in":"query","name":"t","schema":{"type":"string","enum":["*","0","1","2","3"]},"description":"Search scope"},{"in":"query","name":"d[]","schema":{"type":"array","items":{"type":"string"}},"description":"Hadith degree"},{"in":"query","name":"m[]","schema":{"type":"array","items":{"type":"string"}},"description":"Muhadith (narrator) IDs"},{"in":"query","name":"s[]","schema":{"type":"array","items":{"type":"string"}},"description":"Book IDs"},{"in":"query","name":"rawi[]","schema":{"type":"array","items":{"type":"string"}},"description":"Rawi (transmitter) IDs"}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIHadithSearchResponse"}}}}}}},"/site/hadith/search":{"get":{"summary":"Search for hadiths using site data","parameters":[{"in":"query","name":"value","schema":{"type":"string"},"required":true,"description":"Search query for hadith text"},{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number for pagination"},{"in":"query","name":"removehtml","schema":{"type":"boolean"},"description":"Remove HTML tags from results"},{"in":"query","name":"specialist","schema":{"type":"boolean"},"description":"Include specialist hadiths"},{"in":"query","name":"xclude","schema":{"type":"string"},"description":"Words or phrases to exclude from search"},{"in":"query","name":"st","schema":{"type":"string","enum":["w","a","p"]},"description":"Search type (w=all words, a=any word, p=exact phrase)"},{"in":"query","name":"t","schema":{"type":"string","enum":["*","0","1","2","3"]},"description":"Search scope"},{"in":"query","name":"d[]","schema":{"type":"array","items":{"type":"string"}},"description":"Hadith degree"},{"in":"query","name":"m[]","schema":{"type":"array","items":{"type":"string"}},"description":"Muhadith (narrator) IDs"},{"in":"query","name":"s[]","schema":{"type":"array","items":{"type":"string"}},"description":"Book IDs"},{"in":"query","name":"rawi[]","schema":{"type":"array","items":{"type":"string"}},"description":"Rawi (transmitter) IDs"}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteHadithSearchResponse"}}}}}}},"/site/hadith/{id}":{"get":{"summary":"Get a specific hadith by ID from site","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteSingleHadithResponse"}}}}}}},"/site/hadith/similar/{id}":{"get":{"summary":"Get similar hadiths by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteSimilarHadithResponse"}}}}}}},"/site/hadith/alternate/{id}":{"get":{"summary":"Get alternate sahih hadith by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteAlternateHadithResponse"}}}}}}},"/site/hadith/usul/{id}":{"get":{"summary":"Get usul (sources) of a hadith by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteUsulHadithResponse"}}}},"404":{"description":"Hadith not found"}}}},"/site/sharh/search":{"get":{"summary":"Search for sharh (explanations) of hadiths","parameters":[{"in":"query","name":"value","schema":{"type":"string"},"required":true,"description":"Search query for sharh text"},{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number for pagination"},{"in":"query","name":"removehtml","schema":{"type":"boolean"},"description":"Remove HTML tags from results"},{"in":"query","name":"specialist","schema":{"type":"boolean"},"description":"Include specialist hadiths"}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharhSearchResponse"}}}}}}},"/site/sharh/{id}":{"get":{"summary":"Get sharh (explanation) for a hadith by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharhResponse"}}}}}}},"/site/sharh/text/{text}":{"get":{"summary":"Get sharh (explanation) for a hadith by text","parameters":[{"in":"path","name":"text","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharhByTextResponse"}}}}}}},"/site/mohdith/{id}":{"get":{"summary":"Get information about a muhadith by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MohdithResponse"}}}}}}},"/site/book/{id}":{"get":{"summary":"Get information about a book by ID","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponse"}}}}}}},"/data/book":{"get":{"summary":"Get list of all available hadith books","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}},"/data/degree":{"get":{"summary":"Get list of all hadith degrees (authenticity levels)","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}},"/data/methodSearch":{"get":{"summary":"Get method search data for hadith filtering","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}},"/data/mohdith":{"get":{"summary":"Get list of all muhaddithun (hadith narrators)","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}},"/data/rawi":{"get":{"summary":"Get list of all rawi (hadith transmitters)","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}},"/data/zoneSearch":{"get":{"summary":"Get zone search data for geographic hadith filtering","responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataResponse"}}}}}}}},"components":{"schemas":{"SuccessWrapper":{"description":"All responses are wrapped in this structure","type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object"},"data":{"oneOf":[{"type":"object"},{"type":"array"}]}}},"APIHadithSearchResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"currentPageCount":{"type":"integer"},"page":{"type":"integer"},"hasNextPage":{"type":"boolean"},"hasPrevPage":{"type":"boolean"},"removeHTML":{"type":"boolean"},"isCached":{"type":"boolean"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/APIHadithItem"}}}},"APIHadithItem":{"type":"object","properties":{"hadith":{"type":"string"},"rawi":{"type":"string"},"mohdith":{"type":"string"},"book":{"type":"string"},"numberOrPage":{"type":"string"},"grade":{"type":"string"}}},"SiteHadithSearchResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"currentPageCount":{"type":"integer"},"total":{"type":"integer"},"page":{"type":"integer"},"totalPages":{"type":"integer"},"hasNextPage":{"type":"boolean"},"hasPrevPage":{"type":"boolean"},"removeHTML":{"type":"boolean"},"specialist":{"type":"boolean"},"numberOfNonSpecialist":{"type":"integer"},"numberOfSpecialist":{"type":"integer"},"isCached":{"type":"boolean"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/SiteHadithItem"}}}},"HadithCategory":{"type":"object","description":"Thematic category (التصنيف الموضوعي) from dorar.net","properties":{"id":{"type":"string","description":"Category hash ID used in dorar.net URLs (e.g. hadith-category/cat/{id})"},"name":{"type":"string","description":"Category name in Arabic"}}},"SiteHadithItem":{"type":"object","properties":{"hadith":{"type":"string"},"rawi":{"type":"string"},"mohdith":{"type":"string"},"mohdithId":{"type":"string"},"book":{"type":"string"},"bookId":{"type":"string"},"numberOrPage":{"type":"string"},"grade":{"type":"string"},"explainGrade":{"type":"string"},"takhrij":{"type":"string"},"hadithId":{"type":"string"},"categories":{"type":"array","description":"Thematic categories (التصنيف الموضوعي) from dorar.net","items":{"$ref":"#/components/schemas/HadithCategory"}},"hasSimilarHadith":{"type":"boolean"},"hasAlternateHadithSahih":{"type":"boolean"},"hasUsulHadith":{"type":"boolean"},"similarHadithDorar":{"type":"string"},"alternateHadithSahihDorar":{"type":"string"},"usulHadithDorar":{"type":"string"},"urlToGetSimilarHadith":{"type":"string"},"urlToGetAlternateHadithSahih":{"type":"string"},"urlToGetUsulHadith":{"type":"string"},"hasSharhMetadata":{"type":"boolean"},"sharhMetadata":{"type":"object","properties":{"id":{"type":"string"},"isContainSharh":{"type":"boolean"},"urlToGetSharh":{"type":"string"}}}}},"SiteSingleHadithResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"isCached":{"type":"boolean"}}},"data":{"$ref":"#/components/schemas/SiteHadithItem"}}},"SiteAlternateHadithResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"isCached":{"type":"boolean"}}},"data":{"type":"object","properties":{"hadith":{"type":"string"},"rawi":{"type":"string"},"mohdith":{"type":"string"},"mohdithId":{"type":"string"},"book":{"type":"string"},"bookId":{"type":"string"},"numberOrPage":{"type":"string"},"grade":{"type":"string"},"explainGrade":{"type":"string"},"takhrij":{"type":"string"},"hadithId":{"type":"string"},"hasSimilarHadith":{"type":"boolean"},"hasAlternateHadithSahih":{"type":"boolean","description":"Always false for alternate hadith endpoint"},"hasUsulHadith":{"type":"boolean"},"similarHadithDorar":{"type":"string"},"usulHadithDorar":{"type":"string"},"urlToGetSimilarHadith":{"type":"string"},"urlToGetUsulHadith":{"type":"string"},"hasSharhMetadata":{"type":"boolean"},"sharhMetadata":{"type":"object","properties":{"id":{"type":"string"},"isContainSharh":{"type":"boolean"},"urlToGetSharh":{"type":"string"}}}}}}},"SharhResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"isCached":{"type":"boolean"}}},"data":{"$ref":"#/components/schemas/SharhItem"}}},"SharhByTextResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"specialist":{"type":"boolean"},"isCached":{"type":"boolean"}}},"data":{"$ref":"#/components/schemas/SharhItem"}}},"SharhSearchResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"page":{"type":"integer"},"removeHTML":{"type":"boolean"},"specialist":{"type":"boolean"},"isCached":{"type":"boolean"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/SharhItem"}}}},"SharhItem":{"type":"object","properties":{"hadith":{"type":"string"},"rawi":{"type":"string"},"mohdith":{"type":"string"},"book":{"type":"string"},"numberOrPage":{"type":"string"},"grade":{"type":"string"},"takhrij":{"type":"string"},"hasSharhMetadata":{"type":"boolean"},"sharhMetadata":{"type":"object","properties":{"id":{"type":"string"},"isContainSharh":{"type":"boolean"},"urlToGetSharhById":{"type":"string"},"sharh":{"type":"string"}}}}},"MohdithResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"isCached":{"type":"boolean"}}},"data":{"type":"object","properties":{"name":{"type":"string"},"mohdithId":{"type":"string"},"info":{"type":"string"}}}}},"BookResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"isCached":{"type":"boolean"}}},"data":{"type":"object","properties":{"name":{"type":"string"},"bookId":{"type":"string"},"author":{"type":"string"},"reviewer":{"type":"string"},"publisher":{"type":"string"},"edition":{"type":"string"},"editionYear":{"type":"string"}}}}},"SiteSimilarHadithResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"isCached":{"type":"boolean"}}},"data":{"type":"array","items":{"$ref":"#/components/schemas/SiteHadithItem"}}}},"SiteUsulHadithResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"metadata":{"type":"object","properties":{"length":{"type":"integer"},"usulSourcesCount":{"type":"integer"},"isCached":{"type":"boolean"}}},"data":{"type":"object","properties":{"hadith":{"type":"string"},"rawi":{"type":"string"},"mohdith":{"type":"string"},"mohdithId":{"type":"string"},"book":{"type":"string"},"bookId":{"type":"string"},"numberOrPage":{"type":"string"},"grade":{"type":"string"},"explainGrade":{"type":"string"},"takhrij":{"type":"string"},"hadithId":{"type":"string"},"hasSimilarHadith":{"type":"boolean"},"hasAlternateHadithSahih":{"type":"boolean"},"hasUsulHadith":{"type":"boolean"},"similarHadithDorar":{"type":"string"},"alternateHadithSahihDorar":{"type":"string"},"urlToGetSimilarHadith":{"type":"string"},"urlToGetAlternateHadithSahih":{"type":"string"},"hasSharhMetadata":{"type":"boolean"},"sharhMetadata":{"type":"object","properties":{"id":{"type":"string"},"isContainSharh":{"type":"boolean"},"urlToGetSharh":{"type":"string"}}},"usulHadith":{"type":"object","properties":{"sources":{"type":"array","items":{"type":"object","properties":{"source":{"type":"string"},"chain":{"type":"string"},"hadithText":{"type":"string"}}}},"count":{"type":"integer"}}}}}}},"DataResponse":{"type":"object","properties":{"status":{"type":"string","example":"success"},"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}}}}}}}}}